Avoid DB error setting user.address to NULL.
If $lmbperson->adrstreet is not set, do not set $moodleuser->address
to NULL when updating user. If $moodleuser->address is NULL, an error
will occur trying to set user.address to NULL. Also changed behavior
to match $moodleuser->phone1. Instead of $moodleuser->address = '',
do not set $moodleuser->address if $lmbperson->adrstreet is not set
or when ($config->includeaddress && $config->forceaddress) is false.